refactor: service 내부 패키지 정리 (#14)
