+++ src/main/java/com/newfit/reservation/service/equipment/EquipmentService.java
@@ -1,6 +1,6 @@
-package com.newfit.reservation.service;
+package com.newfit.reservation.service.equipment;
 
-import com.newfit.reservation.repository.EquipmentRepository;
+import com.newfit.reservation.repository.equipment.EquipmentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
