+++ src/main/java/com/newfit/reservation/service/location/SiggAreaService.java
@@ -1,6 +1,6 @@
-package com.newfit.reservation.service;
+package com.newfit.reservation.service.location;
 
-import com.newfit.reservation.repository.SiggAreaRepository;
+import com.newfit.reservation.repository.location.SiggAreaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
