+++ src/main/java/com/newfit/reservation/service/reservation/ReservationService.java
@@ -1,7 +1,7 @@
-package com.newfit.reservation.service;
+package com.newfit.reservation.service.reservation;
 
 
-import com.newfit.reservation.repository.ReservationRepository;
+import com.newfit.reservation.repository.reservation.ReservationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
