+++ src/main/java/com/newfit/reservation/service/routine/EquipmentRoutineService.java
@@ -1,6 +1,6 @@
-package com.newfit.reservation.service;
+package com.newfit.reservation.service.routine;
 
-import com.newfit.reservation.repository.EquipmentRoutineRepository;
+import com.newfit.reservation.repository.routine.EquipmentRoutineRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
